requirements-dev.txt
# package management
pip~=22.2
setuptools~=65.3
wheel~=0.37
build~=0.8
twine~=4.0
pip-audit~=2.4

# debugging
django-debug-toolbar~=3.6

# tests
pytest~=7.1
pytest-mock~=3.8
tblib~=1.7
tox~=3.25
coverage~=6.4
django-coverage-plugin~=2.0

# linters
black~=22.6
flake8~=5.0
flake8-annotations-coverage~=0.0.6
flake8-black~=0.3
flake8-bugbear~=22.8
flake8-comprehensions~=3.10
flake8-unused-arguments~=0.0
flake8-django~=1.1

# type checking
mypy~=0.961
pylint~=2.15
pydocstyle~=6.1
pyright~=1.1
types-dateparser~=1.1
types-PyYAML~=6.0
types-requests~=2.28
types-backports~=0.1
pytype~=2022.8.3;python_version<"3.9"
pyre-check~=0.9; platform_system != "Windows"

